+++ TDE/TDE_class.cpp
@@ -10,7 +10,7 @@
 namespace tde
 {
 
-iTDE::iTDE(uint16_t _size, uint16_t _rate, WEIGHTING_FN_TYPE _w_func) :
+iTDE::iTDE(size_t _size, size_t _rate, WEIGHTING_FN_TYPE _w_func) :
     size(_size),
     sample_rate(_rate),
     w_func(_w_func),
@@ -95,7 +95,7 @@ void iTDE::normalize_sum()
     {
         TRACE_EVENT( logger::EVENTS::ERROR, "Inconsistent sizes of amplitude and fourier buffers" );
     }
-    for (uint16_t i = 0; i < size/2+1; ++i)
+    for (size_t i = 0; i < size/2+1; ++i)
     {
         fur_1_2_sum[i] /= update_count;
         ampl1_sum[i] /= update_count;
